LibDl: Move the `dlfcn` implementation to LibC
